// app/types/student.ts
export type StudentType = {
  id: number;
  // regno: number;
  role: string;
  email: string;
  photo: string;
  mobile: string;
  location: string;
  contact: string;
  course: string;
  fullname: string;
  middle_name: string;
  last_name: string;
  sex: string;
  is_active: boolean;
  date_joined: string;
};